refactor(domain) : 도메인 패키지 삭제 후, 각 객체의 역할에 맞는 패키지로 분류
