+++ src/main/java/card/CardSymbol.java
@@ -1,0 +1,18 @@
+package card;
+
+public enum CardSymbol {
+    HEART("하트"),
+    DIAMOND("다이아몬드"),
+    SPADE("스페이드"),
+    CLOVER("클로버");
+
+    private final String name;
+
+    CardSymbol(final String name) {
+        this.name = name;
+    }
+
+    public String getName() {
+        return name;
+    }
+}
