--- src/main/java/domain/ScoreResult/BattleResult.java
@@ -1,17 +1,0 @@
-package domain.ScoreResult;
-
-public enum BattleResult {
-    WIN("승"),
-    LOSE("패"),
-    DRAW("무");
-
-    private final String name;
-
-    BattleResult(final String name) {
-        this.name = name;
-    }
-
-    public String getName() {
-        return name;
-    }
-}
