--- src/test/java/domain/GamblingStatementTest.java
@@ -1,123 +1,0 @@
-package domain;
-
-import domain.ScoreResult.BattleResults;
-import domain.ScoreResult.ScoreBoard;
-import domain.card.GameCardDeck;
-import domain.participant.Dealer;
-import domain.participant.Participant;
-import domain.participant.Participants;
-import domain.participant.Player;
-import java.math.BigDecimal;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-import org.assertj.core.api.Assertions;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-
-public class GamblingStatementTest {
-
-    @Test
-    void 생성_확인() {
-        //given
-        Map<Participant, Money> originGamblingStatement = new HashMap<>();
-
-        //when
-        GamblingStatement gamblingStatement = new GamblingStatement(originGamblingStatement);
-
-        //then
-        Assertions.assertThat(gamblingStatement).isInstanceOf(GamblingStatement.class);
-    }
-
-    @DisplayName("승리시 수익 배팅금액 받기, 플레이어 1승, 플레이어 1승")
-    @Test
-    void 수익_계산1() {
-        //given
-        List<Participant> originParticipants = List.of(new Player("우가"), new Player("베루스"), new Dealer());
-        Participants participants = new Participants(originParticipants);
-
-        Map<Participant, Money> participantMoney = Map.of(
-                participants.getParticipants().getFirst(), new Money("10000"),
-                participants.getParticipants().get(1), new Money("10000"));
-        GamblingStatement gamblingStatement = new GamblingStatement(participantMoney);
-
-        GameCardDeck gameCardDeck = GameCardDeck.generateFullPlayingCard();
-        participants.drawTwoCards(gameCardDeck);
-        participants.drawTwoCards(gameCardDeck);
-
-        //when
-        GamblingStatement profitParticipantMoney = gamblingStatement.calculateProfit(new ScoreBoard(participants));
-
-        Map<Participant, Money> profitMoney = profitParticipantMoney.getGamblingStatement();
-        //then
-        Assertions.assertThat(profitMoney.get(participants.getParticipants().getFirst()).getAmount())
-                .isEqualTo(BigDecimal.valueOf(10000));
-        Assertions.assertThat(profitMoney.get(participants.getParticipants().get(1)).getAmount())
-                .isEqualTo(BigDecimal.valueOf(10000));
-    }
-
-    @DisplayName("무승부시 수익 0원, 플레이어 1무, 플레이어 1무")
-    @Test
-    void 수익_계산2() {
-        //given
-        List<Participant> originParticipants = List.of(new Player("우가"), new Player("베루스"), new Dealer());
-        Participants participants = new Participants(originParticipants);
-        Participants onlyPlayer = participants.findPlayers();
-
-        Participant dealer = participants.findDealer();
-        Participant player1 = onlyPlayer.getParticipants().getFirst();
-        Participant player2 = onlyPlayer.getParticipants().getLast();
-
-
-        Map<Participant, Money> participantMoney = Map.of(
-                participants.getParticipants().getFirst(), new Money("10000"),
-                participants.getParticipants().get(1), new Money("10000"));
-        GamblingStatement gamblingStatement = new GamblingStatement(participantMoney);
-
-        GameCardDeck gameCardDeck = GameCardDeck.generateFullPlayingCard();
-        player1.drawCard(gameCardDeck, 1);
-        player2.drawCard(gameCardDeck, 1);
-        dealer.drawCard(gameCardDeck, 1);
-
-        //when
-        GamblingStatement profitParticipantMoney = gamblingStatement.calculateProfit(new ScoreBoard(participants));
-
-        Map<Participant, Money> profitMoney = profitParticipantMoney.getGamblingStatement();
-
-        //then
-        Assertions.assertThat(profitMoney.get(participants.getParticipants().getFirst()).getAmount())
-                .isEqualTo(BigDecimal.valueOf(0));
-        Assertions.assertThat(profitMoney.get(participants.getParticipants().get(1)).getAmount())
-                .isEqualTo(BigDecimal.valueOf(0));
-    }
-
-    @DisplayName("패배시 수익 -배팅원, 플레이어 1패, 플레이어 1패")
-    @Test
-    void 수익_계산3() {
-        //given
-        List<Participant> originParticipants = List.of(new Player("우가"), new Player("베루스"), new Dealer());
-        Participants participants = new Participants(originParticipants);
-
-        Participant dealer = participants.findDealer();
-
-        Map<Participant, Money> participantMoney = Map.of(
-                participants.getParticipants().getFirst(), new Money("10000"),
-                participants.getParticipants().get(1), new Money("10000"));
-        GamblingStatement gamblingStatement = new GamblingStatement(participantMoney);
-
-        GameCardDeck gameCardDeck = GameCardDeck.generateFullPlayingCard();
-        participants.drawTwoCards(gameCardDeck);
-        dealer.drawCard(gameCardDeck, 4);
-
-        //when
-        GamblingStatement profitParticipantMoney = gamblingStatement.calculateProfit(new ScoreBoard(participants));
-
-        Map<Participant, Money> profitMoney = profitParticipantMoney.getGamblingStatement();
-
-        //then
-        Assertions.assertThat(profitMoney.get(participants.getParticipants().getFirst()).getAmount())
-                .isEqualTo(BigDecimal.valueOf(-10000));
-        Assertions.assertThat(profitMoney.get(participants.getParticipants().get(1)).getAmount())
-                .isEqualTo(BigDecimal.valueOf(-10000));
-    }
-}
